requirements.txt
flask
pytest-playwright
